requirements_yk.txt
absl-py==0.7.1
astor==0.8.0
audioread==2.1.8
backcall==0.1.0
black==21.9b0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
click==8.0.2
colorful==0.5.4
colorthief==0.2.1
cycler==0.10.0
Cython==0.29.10
dataclasses==0.8
decorator==4.4.0
dlib==19.18.0
dominate==2.4.0
easydict==1.9
Eel==0.14.0
entrypoints==0.3
fastrlock==0.6
filelock==3.0.12
flake8==3.7.8
future==0.18.2
gast==0.2.2
google-pasta==0.1.7
graphviz==0.8.4
greenlet==1.0.0
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
importlib-metadata==4.8.1
ipython==7.5.0
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.14.0
joblib==0.13.2
jsonpatch==1.24
jsonpointer==2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
matplotlib==3.1.0
mccabe==0.6.1
msgpack==1.0.2
mypy==0.790
mypy-extensions==0.4.3
networkx==2.3
numba==0.44.1
numpy==1.16.4
olefile==0.46
opencensus==0.7.12
opencv-contrib-python==4.1.1.26
opencv-python==4.1.0.25
opt-einsum==3.3.0
pandas==0.25.1
parso==0.5.0
pathspec==0.9.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==4.3.0
platformdirs==2.4.0
POT==0.4.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pygltflib==1.14.3
Pygments==2.4.2
pynvim==0.4.3
pyparsing==2.4.0
pyssim==0.4
python-dateutil==2.8.0
python-speech-features==0.6
pytz==2019.2
PyWavelets==1.0.3
pyzmq==18.1.0
regex==2021.10.8
requests==2.22.0
resampy==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.0
six==1.12.0
SoundFile==0.10.2
termcolor==1.1.0
toml==0.10.2
tomli==1.2.1
tornado==6.0.3
tqdm==4.19.6
traitlets==4.3.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.25.3
visdom==0.1.8.9
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
whichcraft==0.6.1
wrapt==1.11.1
zipp==3.6.0

# # Install manually
# tensorboard==1.14.0
# tensorflow-estimator==1.14.0
# tensorflow-gpu @ file:///home/chaiyujin/software/tensorflow_gpu-1.14.0%2Bcu102-cp36-cp36m-linux_x86_64.whl
# torch==1.1.0
# torchfile==0.1.0
# torchvision==0.3.0
# mxnet-cu102==1.8.0.post0
